api/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.10.7
joblib==1.2.0
scikit-learn==1.2.2
kafka-python==2.0.2
websockets==10.4
python-dotenv==0.19.0
openai>=1.0.0